created bubble chart and calculations file and started setting up basic framewor for bubble chart
